requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
Pillow==10.1.0
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.26.0
pydantic==2.4.2
python-dotenv==1.0.0